release:  develop changes to main (#148)
* fix: allow empty email in registration input (#59)

* feat: autolink covers for sladers who receive an SMS from EDI (#44)

* feat: isc endpoint to check for permission (#60)

- feat: retrive users with a role, get role by name
- chore: remove unused role usecases

* test: fix tests for getting navigation actions usecase (#53)

* chore: filter roles using filter input struct (#57)

* fix: ensure that returned navigation actions are in the order required (#58)

* chore: update email templates (#62)


* ref: ordering navigation actions using priority (#68)

- chore: add employee navigation actions

* fix: change role filter to search role by role name (#67)

* chore: update registered user welcome message (#70)

fix: return empty list for missing role in role search

* feat: add permission scopes in login response (#69)

* feat: layout community onboarding service deployment (#72)

* feat: layout community onboarding service deployment

[ ] Build and push docker image to GCR
[ ] Deploy the revision to Cloud Run via CI
[ ] Make this workflow stage depend on CI test and lint workflow


* feat: for lint and test workflow to complete and succeed before running deployment workflow


* configure enviroment and job concurrency group


* fix: flagged feature disabled on token refresh (#75)

* refactor purge user (#77)

* feat: add service deployment job (#74)

[ ] Deploy test server to Google Cloud Run
[ ] Supply secrets via Google Secret Manager


* chore: add gitguardian security check and limit deployment to test servers to only develop branch (#78)


* fix: set-env-vars for root collection suffix (#81)


* tests: role management integration and acceptance tests (#63)


* fix: root collection and sentry dns env (#83)


* feat: listing role members in role output and role details in user response (#84)

- assigning multiple roles

* feat: add and save role revocation reason (#73)

chore: roles in admin and agent response

* feat: add AgentProfile to users registered as agents (#76)

* feat: add searching admin by phone (#79)

* feat: add sending temporary pin to users on demand (#54)

* Revert "feat: add sending temporary pin to users on demand (#54)" (#89)

This reverts commit 1ff99b799858580778cdab76e7dfc9218c7d4120.

* chore: add ci schema validation command (#88)


* chore: remove test roles (#90)

* feat: add register user rest endpoint (#105)

* feat: add register user rest endpoint (#115)

* feat: add register user rest endpoint (#116)

* chore: make createdByID an optional attribute when creating a user profile (#124)

* chore: remove email address as required user response (#127)

* fix: add a fix for empty navaction onTapRoutes (#128)

* chore: add consumer registration route (#130)

* fix: add required permission to consumer navigation action (#133)

+++ pkg/onboarding/usecases/roles.go
@@ -25,6 +25,8 @@ type RoleUseCase interface {
 
 	GetAllPermissions(ctx context.Context) ([]*profileutils.Permission, error)
 
+	GetRoleByName(ctx context.Context, name string) (*dto.RoleOutput, error)
+
 	// AddPermissionsToRole adds new scopes to a role
 	AddPermissionsToRole(
 		ctx context.Context,
@@ -38,13 +40,19 @@ type RoleUseCase interface {
 	) (*dto.RoleOutput, error)
 
 	// UpdateRolePermissions replaces the scopes in a role with new updated scopes
-	UpdateRolePermissions(ctx context.Context, input dto.RolePermissionInput) (*dto.RoleOutput, error)
+	UpdateRolePermissions(
+		ctx context.Context,
+		input dto.RolePermissionInput,
+	) (*dto.RoleOutput, error)
 
 	// AssignRole assigns a role to a user
 	AssignRole(ctx context.Context, userID string, roleID string) (bool, error)
 
+	// AssignMultipleRoles assigns multiple roles to a user
+	AssignMultipleRoles(ctx context.Context, userID string, roleIDs []string) (bool, error)
+
 	// RevokeRole removes a role from a user
-	RevokeRole(ctx context.Context, userID string, roleID string) (bool, error)
+	RevokeRole(ctx context.Context, userID, roleID, reason string) (bool, error)
 
 	// ActivateRole marks a role as active
 	ActivateRole(ctx context.Context, roleID string) (*dto.RoleOutput, error)
@@ -54,7 +62,22 @@ type RoleUseCase interface {
 
 	// Check permission checks whether a logged in user with the given UID
 	// is authorized to perform the action specified in the permission
-	CheckPermission(ctx context.Context, uid string, permission profileutils.Permission) (bool, error)
+	CheckPermission(
+		ctx context.Context,
+		uid string,
+		permission profileutils.Permission,
+	) (bool, error)
+
+	// CreateUnauthorizedRole creates a role without performing user authorization
+	// This usecase is useful for creating the initial role in a new environment
+	// and creating the test role used for running integration and acceptance tests
+	CreateUnauthorizedRole(ctx context.Context, input dto.RoleInput) (*dto.RoleOutput, error)
+
+	// UnauthorizedDeleteRole creates a role without performing user authorization
+	// This usecase is useful for cleaning up and removing the test role(s) used for running integration and acceptance tests
+	UnauthorizedDeleteRole(ctx context.Context, roleID string) (bool, error)
+
+	GetRolesByIDs(ctx context.Context, roleIDs []string) ([]*dto.RoleOutput, error)
 }
 
 // RoleUseCaseImpl  represents usecase implementation object
@@ -130,6 +153,54 @@ func (r *RoleUseCaseImpl) CreateRole(
 	return output, nil
 }
 
+// CreateUnauthorizedRole creates a role without performing user authorization
+//
+// This usecase is useful for creating the initial role in a new environment
+// and creating the test role used for running integration and acceptance tests
+// It doesn't check for the users permission
+func (r *RoleUseCaseImpl) CreateUnauthorizedRole(
+	ctx context.Context,
+	input dto.RoleInput,
+) (*dto.RoleOutput, error) {
+	ctx, span := tracer.Start(ctx, "CreateUnauthorizedRole")
+	defer span.End()
+
+	user, err := r.baseExt.GetLoggedInUser(ctx)
+	if err != nil {
+		utils.RecordSpanError(span, err)
+		return nil, err
+	}
+
+	userProfile, err := r.repo.GetUserProfileByUID(ctx, user.UID, false)
+	if err != nil {
+		utils.RecordSpanError(span, err)
+		return nil, err
+	}
+
+	role, err := r.repo.CreateRole(ctx, userProfile.ID, input)
+	if err != nil {
+		utils.RecordSpanError(span, err)
+		return nil, err
+	}
+
+	perms, err := role.Permissions(ctx)
+	if err != nil {
+		utils.RecordSpanError(span, err)
+		return nil, err
+	}
+
+	output := &dto.RoleOutput{
+		ID:          role.ID,
+		Name:        role.Name,
+		Description: role.Description,
+		Active:      role.Active,
+		Scopes:      role.Scopes,
+		Permissions: perms,
+	}
+
+	return output, nil
+}
+
 //GetAllRoles returns a list of all created roles
 func (r *RoleUseCaseImpl) GetAllRoles(ctx context.Context) ([]*dto.RoleOutput, error) {
 	ctx, span := tracer.Start(ctx, "GetAllRoles")
@@ -166,6 +237,13 @@ func (r *RoleUseCaseImpl) GetAllRoles(ctx context.Context) ([]*dto.RoleOutput, e
 			utils.RecordSpanError(span, err)
 			return nil, err
 		}
+
+		users, err := r.repo.GetUserProfilesByRoleID(ctx, role.ID)
+		if err != nil {
+			utils.RecordSpanError(span, err)
+			return nil, err
+		}
+
 		output := &dto.RoleOutput{
 			ID:          role.ID,
 			Name:        role.Name,
@@ -173,6 +251,7 @@ func (r *RoleUseCaseImpl) GetAllRoles(ctx context.Context) ([]*dto.RoleOutput, e
 			Active:      role.Active,
 			Scopes:      role.Scopes,
 			Permissions: perms,
+			Users:       users,
 		}
 		roleOutput = append(roleOutput, output)
 	}
@@ -256,6 +335,31 @@ func (r *RoleUseCaseImpl) DeleteRole(ctx context.Context, roleID string) (bool, 
 		return false, exceptions.RoleNotValid(
 			fmt.Errorf("error: logged in user does not have permissions to delete roles"),
 		)
+	}
+
+	success, err := r.repo.DeleteRole(ctx, roleID)
+	if err != nil {
+		utils.RecordSpanError(span, err)
+		return false, err
+	}
+	return success, nil
+}
+
+// UnauthorizedDeleteRole creates a role without performing user authorization
+//
+// This usecase is useful for cleaning up and removing the test role(s) used for running integration and acceptance tests
+func (r *RoleUseCaseImpl) UnauthorizedDeleteRole(ctx context.Context, roleID string) (bool, error) {
+	ctx, span := tracer.Start(ctx, "UnauthorizedDeleteRole")
+	defer span.End()
+
+	role, err := r.repo.GetRoleByID(ctx, roleID)
+	if err != nil {
+		utils.RecordSpanError(span, err)
+		return false, err
+	}
+
+	if !strings.Contains(strings.ToLower(role.Name), "test") {
+		return false, fmt.Errorf("only test roles can be removed")
 	}
 
 	success, err := r.repo.DeleteRole(ctx, roleID)
@@ -488,6 +592,7 @@ func (r *RoleUseCaseImpl) RevokeRole(
 	ctx context.Context,
 	userID string,
 	roleID string,
+	reason string,
 ) (bool, error) {
 	ctx, span := tracer.Start(ctx, "RevokeRole")
 	defer span.End()
@@ -550,11 +655,30 @@ func (r *RoleUseCaseImpl) RevokeRole(
 		return false, err
 	}
 
+	// revocation input
+	log := dto.RoleRevocationInput{
+		ProfileID: profile.ID,
+		RoleID:    role.ID,
+		Reason:    reason,
+	}
+	// logged in user profile
+	p, err := r.repo.GetUserProfileByUID(ctx, user.UID, false)
+	if err != nil {
+		return false, err
+	}
+	err = r.repo.SaveRoleRevocation(ctx, p.ID, log)
+	if err != nil {
+		return false, err
+	}
+
 	return true, nil
 }
 
 // ActivateRole marks a deactivated role as active and usable
-func (r *RoleUseCaseImpl) ActivateRole(ctx context.Context, roleID string) (*dto.RoleOutput, error) {
+func (r *RoleUseCaseImpl) ActivateRole(
+	ctx context.Context,
+	roleID string,
+) (*dto.RoleOutput, error) {
 	ctx, span := tracer.Start(ctx, "ActivateRole")
 	defer span.End()
 
@@ -617,7 +741,10 @@ func (r *RoleUseCaseImpl) ActivateRole(ctx context.Context, roleID string) (*dto
 }
 
 // DeactivateRole marks a role as inactive and cannot be used
-func (r *RoleUseCaseImpl) DeactivateRole(ctx context.Context, roleID string) (*dto.RoleOutput, error) {
+func (r *RoleUseCaseImpl) DeactivateRole(
+	ctx context.Context,
+	roleID string,
+) (*dto.RoleOutput, error) {
 	ctx, span := tracer.Start(ctx, "DeactivateRole")
 	defer span.End()
 
@@ -680,7 +807,10 @@ func (r *RoleUseCaseImpl) DeactivateRole(ctx context.Context, roleID string) (*d
 }
 
 // UpdateRolePermissions replaces the scopes in a role with new updated scopes
-func (r *RoleUseCaseImpl) UpdateRolePermissions(ctx context.Context, input dto.RolePermissionInput) (*dto.RoleOutput, error) {
+func (r *RoleUseCaseImpl) UpdateRolePermissions(
+	ctx context.Context,
+	input dto.RolePermissionInput,
+) (*dto.RoleOutput, error) {
 	ctx, span := tracer.Start(ctx, "UpdateRolePermissions")
 	defer span.End()
 
@@ -745,9 +875,126 @@ func (r *RoleUseCaseImpl) UpdateRolePermissions(ctx context.Context, input dto.R
 
 // CheckPermission checks whether a logged in user with the given UID
 // is authorized to perform the action specified in the permission
-func (r *RoleUseCaseImpl) CheckPermission(ctx context.Context, uid string, permission profileutils.Permission) (bool, error) {
+func (r *RoleUseCaseImpl) CheckPermission(
+	ctx context.Context,
+	uid string,
+	permission profileutils.Permission,
+) (bool, error) {
 	ctx, span := tracer.Start(ctx, "CheckPermission")
 	defer span.End()
 
 	return r.repo.CheckIfUserHasPermission(ctx, uid, permission)
+}
+
+// GetRoleByName retrieves a role with the matching name
+// Each role has a unique name i.e no duplicate names
+func (r *RoleUseCaseImpl) GetRoleByName(ctx context.Context, name string) (*dto.RoleOutput, error) {
+	ctx, span := tracer.Start(ctx, "GetRoleByName")
+	defer span.End()
+
+	user, err := r.baseExt.GetLoggedInUser(ctx)
+	if err != nil {
+		utils.RecordSpanError(span, err)
+		return nil, err
+	}
+
+	// Check logged in user has the right permissions
+	allowed, err := r.repo.CheckIfUserHasPermission(ctx, user.UID, profileutils.CanViewRole)
+	if err != nil {
+		utils.RecordSpanError(span, err)
+		return nil, err
+	}
+	if !allowed {
+		return nil, exceptions.RoleNotValid(
+			fmt.Errorf("error: logged in user does not have permission to view role"),
+		)
+	}
+
+	role, err := r.repo.GetRoleByName(ctx, name)
+	if err != nil {
+		return nil, err
+	}
+
+	// get permissions
+	perms, err := role.Permissions(ctx)
+	if err != nil {
+		utils.RecordSpanError(span, err)
+		return nil, err
+	}
+
+	output := &dto.RoleOutput{
+		ID:          role.ID,
+		Name:        role.Name,
+		Description: role.Description,
+		Active:      role.Active,
+		Scopes:      role.Scopes,
+		Permissions: perms,
+	}
+
+	return output, nil
+}
+
+// GetRolesByIDs returns the details of roles given the IDs
+func (r RoleUseCaseImpl) GetRolesByIDs(ctx context.Context, roleIDs []string) ([]*dto.RoleOutput, error) {
+	ctx, span := tracer.Start(ctx, "GetRolesByIDs")
+	defer span.End()
+
+	roles, err := r.repo.GetRolesByIDs(ctx, roleIDs)
+	if err != nil {
+		return nil, err
+	}
+
+	roleOutput := []*dto.RoleOutput{}
+	for _, role := range *roles {
+		perms, err := role.Permissions(ctx)
+		if err != nil {
+			utils.RecordSpanError(span, err)
+			return nil, err
+		}
+
+		output := &dto.RoleOutput{
+			ID:          role.ID,
+			Name:        role.Name,
+			Description: role.Description,
+			Active:      role.Active,
+			Scopes:      role.Scopes,
+			Permissions: perms,
+		}
+		roleOutput = append(roleOutput, output)
+	}
+
+	return roleOutput, nil
+}
+
+// AssignMultipleRoles assigns multiple roles to a user
+func (r RoleUseCaseImpl) AssignMultipleRoles(ctx context.Context, userID string, roleIDs []string) (bool, error) {
+	ctx, span := tracer.Start(ctx, "AssignMultipleRoles")
+	defer span.End()
+
+	user, err := r.baseExt.GetLoggedInUser(ctx)
+	if err != nil {
+		utils.RecordSpanError(span, err)
+		return false, err
+	}
+
+	// Check logged in user has the right permissions
+	allowed, err := r.CheckPermission(ctx, user.UID, profileutils.CanAssignRole)
+	if err != nil {
+		utils.RecordSpanError(span, err)
+		return false, err
+	}
+	if !allowed {
+		return false, exceptions.RoleNotValid(
+			fmt.Errorf("error: logged in user does not have permission to view role"),
+		)
+	}
+
+	for _, roleID := range roleIDs {
+		_, err := r.AssignRole(ctx, userID, roleID)
+		if err != nil {
+			return false, err
+		}
+	}
+
+	return true, nil
 }
